Seperated Crucial Types from ones that get exposed
